+++ examples/mcp-github/index.ts
@@ -1,29 +1,24 @@
 #!/usr/bin/env npx -y bun
 
 import assert from "node:assert";
-import { AIAgent, ExecutionEngine, MCPAgent, getMessage } from "@aigne/core";
-import { OpenAIChatModel } from "@aigne/core/models/openai-chat-model.js";
+import { runChatLoopInTerminal } from "@aigne/cli/utils/run-chat-loop.js";
+import { AIAgent, ExecutionEngine, MCPAgent } from "@aigne/core";
+import { loadModel } from "@aigne/core/loader/index.js";
 import { logger } from "@aigne/core/utils/logger.js";
-import { runChatLoopInTerminal } from "@aigne/core/utils/run-chat-loop.js";
 
-const { OPENAI_API_KEY, GITHUB_PERSONAL_ACCESS_TOKEN } = process.env;
-assert(OPENAI_API_KEY, "Please set the OPENAI_API_KEY environment variable");
-assert(
-  GITHUB_PERSONAL_ACCESS_TOKEN,
-  "Please set the GITHUB_PERSONAL_ACCESS_TOKEN environment variable",
-);
+const { GITHUB_TOKEN } = process.env;
+
+assert(GITHUB_TOKEN, "Please set the GITHUB_TOKEN environment variable");
 
 logger.enable(`aigne:mcp,${process.env.DEBUG}`);
 
-const model = new OpenAIChatModel({
-  apiKey: OPENAI_API_KEY,
-});
+const model = await loadModel();
 
 const github = await MCPAgent.from({
   command: "npx",
   args: ["-y", "@modelcontextprotocol/server-github"],
   env: {
-    GITHUB_PERSONAL_ACCESS_TOKEN,
+    GITHUB_TOKEN,
   },
 });
 
@@ -54,7 +49,6 @@ await runChatLoopInTerminal(userAgent, {
   welcome:
     "Hello! I'm a chatbot that can help you interact with GitHub. Try asking me a question about GitHub repositories!",
   defaultQuestion: "Search for repositories related to 'aigne-framework'",
-  onResponse: (response) => console.log(getMessage(response)),
 });
 
 process.exit(0);