+++ examples/mcp-sqlite/index.ts
@@ -1,20 +1,14 @@
 #!/usr/bin/env npx -y bun
 
-import assert from "node:assert";
 import { join } from "node:path";
-import { AIAgent, ExecutionEngine, MCPAgent, PromptBuilder, getMessage } from "@aigne/core";
-import { OpenAIChatModel } from "@aigne/core/models/openai-chat-model.js";
+import { runChatLoopInTerminal } from "@aigne/cli/utils/run-chat-loop.js";
+import { AIAgent, ExecutionEngine, MCPAgent, PromptBuilder } from "@aigne/core";
+import { loadModel } from "@aigne/core/loader/index.js";
 import { logger } from "@aigne/core/utils/logger.js";
-import { runChatLoopInTerminal } from "@aigne/core/utils/run-chat-loop.js";
-
-const { OPENAI_API_KEY } = process.env;
-assert(OPENAI_API_KEY, "Please set the OPENAI_API_KEY environment variable");
 
 logger.enable(`aigne:mcp,${process.env.DEBUG}`);
 
-const model = new OpenAIChatModel({
-  apiKey: OPENAI_API_KEY,
-});
+const model = await loadModel();
 
 const sqlite = await MCPAgent.from({
   command: "uvx",
@@ -43,7 +37,6 @@ const userAgent = engine.call(agent);
 
 await runChatLoopInTerminal(userAgent, {
   initialCall: {},
-  onResponse: (response) => console.log(getMessage(response)),
 });
 
 process.exit(0);